Update pointnotice division
