+++ data/src/main/java/com/dsm/dms/data/repository/NoticeRepositoryImpl.kt
@@ -5,7 +5,6 @@ import com.dsm.dms.data.entity.toDataEntity
 import com.dsm.dms.data.entity.toDbEntity
 import com.dsm.dms.data.entity.toEntity
 import com.dsm.dms.domain.entity.Notice
-import com.dsm.dms.domain.entity.Point
 import com.dsm.dms.domain.repository.NoticeRepository
 import io.reactivex.Single
 
@@ -33,23 +32,4 @@ class NoticeRepositoryImpl(
     override fun deleteLocalNotice(id: Int) =
         dataSource.deleteLocalNotice(id)
 
-    override fun getRemotePointNotice(): Single<List<Point>> =
-        dataSource.getRemotePointNotice().map { response ->
-            response.point.map { data ->
-                data.toEntity()
-            }
-        }
-
-    override fun getLocalPointNotice(): List<Point>? =
-        dataSource.getLocalPointNotice()?.map { it.toDataEntity().toEntity() }
-
-    override fun saveLocalPointNotice(vararg point: Point) =
-        dataSource.saveLocalPointNotice(
-            *point.map {
-                it.toDataEntity().toDbEntity()
-            }.toTypedArray()
-        )
-
-    override fun deleteLocalPointNotice(name: String) =
-        dataSource.deleteLocalPointNotice(name)
 }
