+++ domain/src/main/java/com/dsm/dms/domain/repository/PointNoticeRepository.kt
@@ -1,0 +1,17 @@
+package com.dsm.dms.domain.repository
+
+import com.dsm.dms.domain.entity.Point
+import io.reactivex.Single
+
+
+interface PointNoticeRepository {
+
+    fun getRemotePointNotice(): Single<List<Point>>
+
+    fun getLocalPointNotice(): List<Point>?
+
+    fun saveLocalPointNotice(vararg point: Point)
+
+    fun deleteLocalPointNotice(name: String)
+
+}
