+++ domain/src/main/java/com/dsm/dms/domain/service/NoticeService.kt
@@ -2,14 +2,11 @@ package com.dsm.dms.domain.service
 
 import com.dsm.dms.domain.base.Result
 import com.dsm.dms.domain.entity.Notice
-import com.dsm.dms.domain.entity.Point
 import io.reactivex.Single
 
 
 interface NoticeService {
 
     fun getNotice(): Single<Result<List<Notice>>>
-
-    fun getPointNotice(): Single<Result<List<Point>>>
 
 }
