+++ domain/src/main/java/com/dsm/dms/domain/service/NoticeServiceImpl.kt
@@ -3,7 +3,6 @@ package com.dsm.dms.domain.service
 import com.dsm.dms.domain.base.ErrorHandler
 import com.dsm.dms.domain.base.Result
 import com.dsm.dms.domain.entity.Notice
-import com.dsm.dms.domain.entity.Point
 import com.dsm.dms.domain.repository.NoticeRepository
 import com.dsm.dms.domain.toResult
 import io.reactivex.Single
@@ -20,14 +19,6 @@ class NoticeServiceImpl(
                 handler = handler,
                 localData = repository.getLocalNotice(),
                 saveLocalFun = { T -> repository.saveLocalNotice(*T.toTypedArray()) }
-            )
-
-    override fun getPointNotice(): Single<Result<List<Point>>> =
-        repository.getRemotePointNotice()
-            .toResult(
-                handler = handler,
-                localData = repository.getLocalPointNotice(),
-                saveLocalFun = { T -> repository.saveLocalPointNotice(*T.toTypedArray()) }
             )
 
 }
